refactor: update user management and API integration
- Refactored user management component to utilize a new user status update method.
- Removed deprecated user status handling and replaced it with a more robust implementation.
- Introduced a new API module for handling requests, improving code organization and maintainability.
- Updated imports across services to reflect the new API structure.
